Fix egui viewport dimensions when using a non-integer pixel scaling factor
Prior to this fix, was casting pixels_per_point to a usize, so
non-integer scaling factors (for instance, 1.5) resulted in rendering
the gui with the wrong viewport dimensions. This bug could be reproduced
by running the lighting example with a display scale of 150%.

The main fix here is to ensure the cast happens after applying the scaling
factor, and to make sure the scaling factor multiplication happens in f32.

A more subtle change is that we explicitly round the result to the nearest
integer. This is to reduce the potential error in cases where the
multiplication does not produce a whole number.