Deprecate update_transformations()
A majority of this 'update' function needed to be extracted to allow being called externally without a new set of transformations. As `update_buffers()` is now the new function for the internal updating of the struct, the semantics for a new slice of transforms to apply should be `set`.

Note also that `update_buffers()` is public so that the new `add` functions can be called procedurally without the overhead of updating each time, then update can be called externally afterward.