Texture transform (#164)
* TextureRegion struct, InstancedModel properties for it. 'subt' program attribute

Add subtexture support to instanced models.

Provides uv offset and scale to shader per instance, allowing texture adjustment. Useful as simple uv scrolling or scaling per instance or atlasing multiple textures for a single instance.

* Revert "Fix error in light iterator"

This reverts commit 1119ddb46b885deaa22d3aa3bb11286c92020e97.

* Transformation and buffer cleanup in InstancedModel

Remove redundant aabb. Split buffer processing from transform update and also purpose for subtexture updates.

* InstancedModel docs

* Pass subtexture buffer only when needed.

* Semantic consistency.

* cargo fmt

* Use basic transformation property if transforms array is empty.

* Deprecate update_transformations()

A majority of this 'update' function needed to be extracted to allow being called externally without a new set of transformations. As `update_buffers()` is now the new function for the internal updating of the struct, the semantics for a new slice of transforms to apply should be `set`.

Note also that `update_buffers()` is public so that the new `add` functions can be called procedurally without the overhead of updating each time, then update can be called externally afterward.

* TextureRegion -> TextureTransform, use some gl semantics.

* Instances instead of separate transforms. Indexed pushing and updating.

* Restore update_buffers() calls in instance handling until mutability stuff is worked out.

* Revert "Revert "Fix error in light iterator""

This reverts commit 3c167590f848e68d75b64c31656d754ef53e3f63.

* Fix merge problem

* Default for texture transform

* Take ref slice as input instead of Vec

* Clean-up in instanced model

* Avoid instance_count

* Model/InstancedModel support texture transform

* Get/set texture transform

* Minor clean-up

* More clean-up

* Yet more clean-up

* Fix normals example

Co-authored-by: Ranzear <Ranzear@gmail.com>