Fix canvas size on fractional device pixel ratios
This commit aims to fix some rounding errors that occur when having a
non-integer device pixel ratio. This made it impossible to render
pixel-perfect in some cases, which would be required in things like
pixelart games or maybe other cases like UI rendering.

I tried to keep the changes as minimal as possible, but only with these
changes I was able to render pixel-perfect textures, where one pixel in
the texture exactly matches a pixel in the final canvas:

- Allow the size given in the CSS styling to be fractional, as this is
  intended behavior. For example, if the canvas is 200 native pixels
  wide (so the HTML attribute says width="200"), with a device pixel
  ratio of 1.5 the CSS width should be:
  (200/1.5)px = 133.333333px
- Made sure width & height passed to the viewport in the FrameInput is
  the actual size in native pixels. Previously, the original native
  pixels (returned by `canvas.width()` and `canvas.height()`) were
  first converted to account for the device pixel ratio and then
  rounded, as this is expected for the window_width and window_height
  attributes in the FrameInput. However, then these values were
  extrapolated again to get the viewport width + height, which
  resulted in rounding errors to make the viewport not match the actual
  size in native pixels.