Make all structs Send and Sync.
This replaces all occurences of `std::rc::Rc` with `std::sync::Arc`
and all occurences of `std::cell::RefCell` with `std::sync::RwLock`.

This makes all structs Send and Sync, when the `glutin` features
is not enabled, allowing them to be passed along when the OpenGL
context is passed from one thread to another.