Optimise full_screen_draw
Get rid of the vertex buffer entirely, in favour of indexing vertex positions from an array within the vertex shader.

Before this patch full_screen_draw is taking 18% of my frametime (mostly in creating and destroying the vertex buffer many times per from).

After the patch it is reduced to 5%. Note that my app is very heavy on post-process effects - I expect the majority of apps will not see quite such a major difference.